requirements.txt
fastapi
mangum
llama-cpp-python
awslambdaric